+++ examples/trajectories/example_pulseq.py
@@ -11,7 +11,7 @@ import numpy as np
 from mrinufft.display import display_3D_trajectory
 from mrinufft.trajectories import initialize_2D_spiral, stack
 
-from mrinufft.io.pulseq import pulseq_gre, read_pulseq_traj
+from mrinufft.io import pulseq_gre, read_pulseq_traj
 import matplotlib.pyplot as plt
 
 # %%
